+++ WebShop/wwwroot/AreaAdmin/css/custom.css
@@ -14,141 +14,137 @@
     =============================================*/
 body {
     font-family: 'Open Sans', sans-serif;
-    line-height:30px;
+    line-height: 30px;
 }
 
- #wrapper {
+#wrapper {
     width: 100%;
-    margin-top:50px;
+    margin-top: 50px;
 }
 
 #page-wrapper {
     padding: 15px 15px;
     min-height: 600px;
-    background:#F3F3F3;
-   
+    background: #F3F3F3;
 }
+
 #page-inner {
-    width:100%;
-    margin:10px 20px 10px 0px;
-    background-color:#fff!important;
-    padding:10px;
-    min-height:1050px;
+    width: 100%;
+    margin: 10px 20px 10px 0px;
+    background-color: #fff !important;
+    padding: 10px;
+    min-height: 1050px;
 }
+
 .pad-top {
-    padding-top:15px;
+    padding-top: 15px;
 }
 /*==============================================
     MENU STYLES    
     =============================================*/
 
 .sidebar-collapse .nav {
-	padding:0;
-    padding-top:50px;
+    padding: 0;
+    padding-top: 50px;
 }
 
 .sidebar-collapse > .nav > li > a {
-	padding:15px 10px;
+    padding: 15px 10px;
     color: #214761;
-font-weight: 900;
-font-size: 16px;
+    font-weight: 900;
+    font-size: 16px;
 }
+
 .sidebar-collapse > .nav > li {
-	border-bottom:1px solid #e1e1e1;
+    border-bottom: 1px solid #e1e1e1;
 }
+
 .sidebar-collapse .nav > li > a:hover,
 .sidebar-collapse .nav > li > a:focus {
-	
-	background:#F3F3F3;
-	outline:0;
+    background: #F3F3F3;
+    outline: 0;
 }
 
 .navbar-side {
-	border:none;
-	background-color: #fff;
-	
+    border: none;
+    background-color: #fff;
 }
-.navbar-inverse {
 
-background-color: #214761;
-border-color: #F0F0F0;
-min-height: 80px;
+.navbar-inverse {
+    background-color: #214761;
+    border-color: #F0F0F0;
+    min-height: 80px;
 }
 
 
 .nav > li > a > i {
-    margin-right:10px;
+    margin-right: 10px;
 }
+
 .logout-spn {
-    float:right;
-    color:#fff;
-    font-size:25px;
-    font-weight:800;
-    margin:20px;
+    float: right;
+    color: #fff;
+    font-size: 25px;
+    font-weight: 800;
+    margin: 20px;
     margin: 10px 25px 10px 50px;
 }
+
     .logout-spn a {
-        text-decoration:none;
+        text-decoration: none;
     }
+
 .active-link {
-    background-color:#f3f3f3;
+    background-color: #f3f3f3;
 }
-    
+
 /*==============================================
     DASHBOARD STYLES    
     =============================================*/
 .div-square {
-    padding:5px;
-    border:3px double #e1e1e1;
-    -webkit-border-radius:8px;
-   -moz-border-radius:8px;
-    border-radius:8px;
-    margin:5px;
-
+    padding: 5px;
+    border: 3px double #e1e1e1;
+    -webkit-border-radius: 8px;
+    -moz-border-radius: 8px;
+    border-radius: 8px;
+    margin: 5px;
 }
 
-.div-square> a,.div-square> a:hover {
-    color:#808080;
-     text-decoration:none;
-}
+    .div-square > a, .div-square > a:hover {
+        color: #808080;
+        text-decoration: none;
+    }
 /*==============================================
     FOOTER STYLES     
     =============================================*/
 .footer {
-    background-color:#214761;
-    width:100%;
-    color:#fff;
-    padding:20px 50px 20px 50px;
-     padding-left:20px;
+    background-color: #214761;
+    width: 100%;
+    color: #fff;
+    padding: 20px 50px 20px 50px;
+    padding-left: 20px;
 }
 
 
-    .footer > a, .footer>a:hover {
-       
-        color:#fff;
+    .footer > a, .footer > a:hover {
+        color: #fff;
     }
 
 /*==============================================
     MEDIA QUERIES     
     =============================================*/
- 
- @media(min-width:768px) {
-     #page-wrapper{
-               margin: 0 0 0 260px;
+
+@media(min-width:768px) {
+    #page-wrapper {
+        margin: 0 0 0 260px;
         padding: 15px 30px;
         min-height: 1050px;
-		
     }
-	
-	
+
+
     .navbar-side {
         z-index: 1;
         position: absolute;
         width: 260px;
-       
     }
-
- 
 }
-
-
